Import gpg key only on node 0

+++ setup_gpg.sh
@@ -2,10 +2,12 @@
 
 # Important to not use -x here, or secrets will be exposed...
 
-if [[ $CIRCLE_BRANCH == "master" ]]; then
-    echo -e "use-agent\ndefault-key cljsjs.robot@gmail.com\npassphrase $GPG_PASSPHRASE" >> ~/.gnupg/gpg.conf
-    # CircleCI doesn't support newlines in vars so newlines are coded as "|"
-    echo -e $GPG_KEY | tr "|" "\n" > gpg.key
-    gpg --allow-secret-key-import --import gpg.key
-    rm gpg.key
+if [[ $CIRCLE_NODE_TOTAL < 2 ]] || [[ $CIRCLE_NODE_INDEX == 0 ]]; then
+    if [[ $CIRCLE_BRANCH == "master" ]]; then
+        echo -e "use-agent\ndefault-key cljsjs.robot@gmail.com\npassphrase $GPG_PASSPHRASE" >> ~/.gnupg/gpg.conf
+        # CircleCI doesn't support newlines in vars so newlines are coded as "|"
+        echo -e $GPG_KEY | tr "|" "\n" > gpg.key
+        gpg --allow-secret-key-import --import gpg.key
+        rm gpg.key
+    fi
 fi
